requirements.txt
numpy==1.22.0
pandas==1.4.1
Pillow==9.0.1
streamlit==1.7.0
yfinance==0.1.70
